requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.42.1
crispy-bootstrap5==2024.10
dj-database-url==0.5.0
Django==5.1.5
django-allauth==65.3.1
django-cloudinary-storage==0.3.0
django-crispy-forms==2.3
idna==3.10
pillow==11.1.0
psycopg2==2.9.10
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.8.2
beautifulsoup4==4.10.0
requests==2.26.0